Python/Advanced Python/Pacman_DQL/requirements.txt
ale_py==0.7.3
gym==0.21.0
imageio==2.9.0
ipython==8.0.1
matplotlib==3.4.3
numpy==1.21.0
Pillow==9.0.0
PyVirtualDisplay==2.2
tensorflow==2.7.0
tf_agents==0.11.0
